prevent package from loading all dependencies unless called from cli
